Update stripe_waterfall.sql
Fixed aggregation issues by leveraging a grouping_key which is a just a COALESCE of sfdc_account_id and stripe_customer_id. Allows for dynamic partitioning based on whether the stripe customer is tied to a SFDC account yet. 

Also added in a concatenated stripe_customer_id so we can see which accounts have multiple stripe customer IDs and what the Ids are for quick verification in stripe